There are two major alterations in this present version:
1) The code has a corrected psuedo-random number generator (RNG).
This correction was simply a matter of including a 'save' to save variables that were needed to seed subsuequent calls of the RNG.

2)The loop in cmorc.f that includes the radiative transfer has been successfully parallelized.
The time saving benefits are only significant if a non-integer valued IR opacity law is used.
Otherwise, the overhead costs evidently lead to negligible savings.

Note: Results computed with parallelization turned on are NOT exactly reproducable.
This is consequence of the random number generator being called by each cores.  When run
serially, the RNG sequence is deterministic, with a known starting value.
When run in serial, each core is seeded individually by whatever the currently saved value
of the seed is (updated each time cran.f is run). Since this value depends on the sequence
and the precise timing, the results are effectively non-deterministic on the level of noise.
So far, with runs of just 5 days, the differences are insignificant, usually less than 1%.