In this substantial update, the user interface for including aerosols was developed. The fort.7 file now includes a revised radiation namelist and a new cloud specific namelist. An additional common block was added to pass these cloud variables along.
Clouds and/or hazes are specified by a total integrated visible optical thickness,
a multiplicitave factor to relate the optical thickness in the longwave channel,
single scattering albedos and asymmetry parameters for both channels, top and bottom pressures,
and a power law to determine the relative aerosol scale height relative to the pressure scale height.
All these parameters are specified in the fort.7 input file, which means the user doesn't have to
open the hood, modify the code, and recompile.

A new subroutine rmakeclouds.f was written to generate an array specifying the cloud distribution
as determined by the input of the user in fort.7.  The cloud information is written to the new file,
fort.77, that is written for user readability.

Current cloud models supported are:
1) 'Kepler7b'--dayside distribution (from Munoz & Isaak, 2015)
2) 'Nightside'--covers the nightside with clouds
3) 'Global'--covers the entire planet with clouds